Restyle KPI Table and cleanup code
